Bring EvaluateConfigurations and DslEntry to internal
