+++ main/src/test/scala/sbt/internal/parser/CommentedXmlSpec.scala
@@ -1,4 +1,4 @@
-package sbt.internals.parser
+package sbt.internal.parser
 
 class CommentedXmlSpec extends CheckIfParsedSpec {
