+++ main/src/test/scala/sbt/internal/parser/EmbeddedXmlSpec.scala
@@ -1,4 +1,4 @@
-package sbt.internals.parser
+package sbt.internal.parser
 
 import sbt.internal.util.MessageOnlyException
